+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.1.0
 
-* Added the "addValueListener" parameter to listen to its Menu Value.
+* Added the `addValueListener` parameter to listen to the `DraggableMenu`'s Menu Value.
 
-* Added the "Soft Modern UI" as a new UI
+* Added the `Soft Modern UI` as a new UI.
+
+* Changed the return type of the `open` and the `openReplacement` methods of the `DraggableMenu`.
+
+* Added the `blockMenuClosing` parameter to the `DraggableMenu`.